+++ src/class.rs
@@ -14,7 +14,6 @@ use rb_sys::{
 };
 
 use crate::{
-    debug_assert_value,
     error::{protect, Error},
     exception,
     into_value::{ArgList, IntoValue},