+++ src/r_bignum.rs
@@ -10,7 +10,6 @@ use rb_sys::{
 };
 
 use crate::{
-    debug_assert_value,
     error::{protect, Error},
     exception,
     integer::{Integer, IntegerType},