+++ src/r_struct.rs
@@ -17,7 +17,6 @@ use rb_sys::{
 
 use crate::{
     class::RClass,
-    debug_assert_value,
     error::{protect, Error},
     exception,
     into_value::IntoValue,