+++ src/symbol.rs
@@ -5,7 +5,6 @@ use rb_sys::{
 };
 
 use crate::{
-    debug_assert_value,
     encoding::EncodingCapable,
     error::{protect, Error},
     exception,