+++ src/try_convert.rs
@@ -5,7 +5,6 @@ use rb_sys::{rb_get_path, rb_num2dbl};
 #[cfg(ruby_use_flonum)]
 use crate::value::Flonum;
 use crate::{
-    debug_assert_value,
     error::{protect, Error},
     exception,
     integer::Integer,